Move generic helper testing functions to separate file.
